Extract single file views to file
