+++ packages/expo-ui/ios/TextView.swift
@@ -18,7 +18,6 @@ internal final class TextViewProps: UIBaseViewProps {
 internal struct TextView: ExpoSwiftUI.View {
   @ObservedObject var props: TextViewProps
 
-
   var body: some View {
     let hasDeprecatedFontProps = props.weight != nil || props.design != nil || props.size != nil
